Remove treatment flag with a TODO in sentences_preprocessed (Recidiviz/recidiviz-data#16118)
## Description of the change

Per the discussion in this [slack
thread](https://recidiviz.slack.com/archives/C02UDHE0DQR/p1666207373229959)
our current methodology for flagging treatment flags seems to be flawed,
partially because of bugs and partially because MO is also uncertain
about the right path for flagging 120-day sentences. This PR removes
TREATMENT as a value from `sentence sub type` to avoid confusion in
downstream views/dashboards and adds a TODO to add it back in when we
know the right way to flag these sentences.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 05c1279fd56d1c69456fc7d0f324764e9bd1b9d3