[Validation] Make sameness view check run logic in BQ (Recidiviz/recidiviz-data#16141)
## Description of the change

Some of Tennessee's external validations are really slow, especially
ones that compare supervision populations on an individual level. It
looks like the existing BigQuery queries used to run the validations,
while they process a decent amount of data, don't actually take all that
long. The actual time is actually just trying to send all the result
rows back to the client, which takes a long time, uses a lot of CPU, and
seems to even be OOMing our appengine instances some times.

We only need some summary statistics when running the validation, not
all of the results, so this adds another layer of BQ queries to
calculate those summary statistics and just send those back to the
client. Running this locally on my machine got the longest validation,
`supervision_population_person_level_external_comparison_matching_people_district`,
to run in 32.83 seconds. I wasn't patient enough to wait for it to
complete before this change, but in appengine it looked like it was
taking well over an hour most of the time.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

N/A

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: df8b22dd4046bfa92f0914c5bf195cda18296e1e