[Justice Counts] run migration to update metric keys again (Recidiviz/recidiviz-data#16247)
## Description of the change

This PR is in response to [this
error](https://console.cloud.google.com/errors/detail/CJrf1sbxuYPK1gE;time=P1D?project=recidiviz-staging)
in the staging logs where there is a key error related to a metric
definition key that included disaggregation information. When I
inspected the staging db, there seemed to be datapoints with both the
old metric definition keys and new metric definition keys. The
datapoints with the old keys belong to a report that was last updated on
9/29. I believe the migration was run around the same time and
potentially the same day (the PR out for the original migration was
closed on 9/29).

I searched the logs to see any errors associated with a migration and I
couldn't find any. I'm assuming that it failed silently and only updated
the db partially. That said, please let me know if you have any other
leads that I should check out to debug this any further.

## Related issues

Contributes to Recidiviz/recidiviz-data#16156

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e0fc3ff498bbd1f9483c534dfde2762f9a0ad72b