Tool to set clustering fields on a bigquery table (Recidiviz/recidiviz-data#16095)
## Description of the change

This adds a tool to let us set clustering fields on a given table. This
is needed to be able to migrate all of our raw data tables to be
clustered by `file_id`.

I ran this on `us_tn_raw_data.ContactNoteComment` in both staging and
prod to verify that it works correctly.

Right now this just does one file, but I think to do the rest of the
migration we should:

1. Make it operate on one state at a time
2. Run it for each state in staging, making sure the queues are paused
so nothing is writing/reading from the tables
3. Verify it worked, didn't delete data, imports still work fine, delete
queries are free now (!!)
4. Do the same in prod.
5. Update
`recidiviz/ingest/direct/raw_data/direct_ingest_raw_data_table_manager.py`
to set `clustering_fields=["file_id"]` for new tables

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#15952, Recidiviz/recidiviz-data#15201

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 678162464a52058f09199fe01b05e49e629400fb