remove all references to description field passed through as a query kwarg (Recidiviz/recidiviz-data#16275)
## Description of the change

The description is already passed to `BigQueryViewBuilder` via a
parameter, so it does not need to be passed as a query kwarg. This 1st
PR removes all references to description as a query kwarg in all views.
It does not seem like there are any meaningful tests to be added since
the behavior of the views does not change. A few tests are modified
because they explicitly check for the presence of the description in the
query.

A follow up 2nd PR will modify `BigQueryViewBuilder` to not push it
through at all.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

1st of 2 PR's for Recidiviz/recidiviz-data#14545

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7ee0197b85f93ac66fbbdfe99646f2d33c168bea