Record runtime of each validation (Recidiviz/recidiviz-data#16202)
## Description of the change

Updates the validation manager to time how long each validation takes to
run and both record this in the BigQuery table and add it to the log.
This will make it slightly easier to debug which validations are taking
the longest in the future (this time it was pretty obvious lol).

To do this I factored out `_timed_func` that we use in the dag walker to
be a common utility, and went ahead and updated types on these wrapper
functions now that I realized we can pull in `ParamSpec` from
`typing_extensions` and mypy understands it.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

follow up to TN validation timeouts

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a0c6c4e42a584311dcd6349c650e23536863e830