Support sandbox prefixes for normalized state (Recidiviz/recidiviz-data#15528)
## Description of the change

Updates the step that unions the state specific normalized datasets and
copies the non-normalized entities from state into `normalized_state` to
work on sandboxes.

You can now specify a sandbox prefix that it will use as the prefix for
`state` and `normalized_state`, and also tell it whether to pull the
normalized entities from the sandbox or the regular datasets on a state
code by state code basis. (Although maybe we should just be skipping
other states instead?)

Did this to unblock @gheidkamp11 trying to validate the MO rerun in
secondary, but should prove useful for the "dag epic" as well.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

start of Recidiviz/recidiviz-data#15526

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 492ec0a6958c6f77a7c03f8d149e16dd6287108c