Fix small bug in correctional level deduping (Recidiviz/recidiviz-data#16311)
## Description of the change

While testing output from updates to end_date logic in sessions, I
noticed a small unrelated discrepancy. This is related to the new
"INVESTIGATION" `compartment_level_1` value, which can have a
supervision level value, but is not getting deduplicated properly due to
the logic in that join condition.

I removed the restriction that the correctional level deduplication
priority only gets joined in from "SUPERVISION" and
"SUPERVISION_OUT_OF_STATE" which means that it is now being applied to
"INVESTIGATION" as well. Given that the correctional level value is only
population for these 3 `compartment_level_1` values, removing that from
the join condition entirely seemed to make most sense to me.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 98c987d96496d2179d356a59f8dada380a25196d